Add: enable a5 onboard support for L3-L2 orch comm (#1277)
- Enable a5 onboard L3-L2 comm through the common DeviceRunner path.
- Move the queue and stream examples under examples/workers/l3 with four-platform coverage.
- Drop the a5-only ST mirror now that examples cover a5 onboard.
