+++ examples/workers/l3/l3_l2_orch_comm_stream/test_l3_l2_orch_comm_stream.py
@@ -7,7 +7,10 @@
 # INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
 # See LICENSE in the root of the software repository for the full text of the License.
 # -----------------------------------------------------------------------------------------------------------
-"""a5sim-only L3-L2 in-flight communication regression; a5 onboard is unsupported."""
+"""Closed-loop L3-L2 in-flight orchestration communication stream demo.
+
+This file is both a runnable example and a pytest scene-test entry.
+"""
 
 from __future__ import annotations
 
@@ -103,16 +106,13 @@ def _assert_output_matches(output_tensor, expected: list[float]) -> None:
         assert abs(got - want) <= 1e-5, f"output[{i}] expected {want}, got {got}"
 
 
-@pytest.mark.platforms(["a5sim"])
-@pytest.mark.device_count(1)
-@pytest.mark.runtime(_RUNTIME)
-def test_closed_loop_payload_signal_path_while_l2_task_is_in_flight(st_platform, st_device_ids):
-    chip_callable = _build_chip_callable(st_platform)
+def run_closed_loop_stream(platform: str, device_id: int) -> None:
+    chip_callable = _build_chip_callable(platform)
     worker = Worker(
         level=3,
-        device_ids=[int(st_device_ids[0])],
+        device_ids=[int(device_id)],
         num_sub_workers=0,
-        platform=st_platform,
+        platform=platform,
         runtime=_RUNTIME,
     )
     handle = worker.register(chip_callable)
@@ -166,3 +166,10 @@ def test_closed_loop_payload_signal_path_while_l2_task_is_in_flight(st_platform,
         worker.run(orch, args=None, config=config)
     finally:
         worker.close()
+
+
+@pytest.mark.platforms(["a2a3sim", "a2a3", "a5sim", "a5"])
+@pytest.mark.device_count(1)
+@pytest.mark.runtime("tensormap_and_ringbuffer")
+def test_l3_l2_orch_comm_stream(st_platform, st_device_ids):
+    run_closed_loop_stream(st_platform, int(st_device_ids[0]))
