--- tests/st/a5/tensormap_and_ringbuffer/l3_l2_orch_comm/kernels/aiv/kernel_l3_l2_transform.cpp
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) PyPTO Contributors.
- * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
- * CANN Open Software License Agreement Version 2.0 (the "License").
- * Please refer to the License for details. You may not use this file except in compliance with the License.
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
- * See LICENSE in the root of the software repository for the full text of the License.
- * -----------------------------------------------------------------------------------------------------------
- */
-
-#include <cstdint>
-#include <pto/pto-inst.hpp>
-
-#include "pipe_sync.h"
-#include "tensor.h"  // NOLINT(build/include_subdir)
-
-// NOLINTNEXTLINE(build/namespaces)
-using namespace pto;
-
-#ifndef __gm__
-#define __gm__
-#endif
-
-#ifndef __aicore__
-#define __aicore__ [aicore]  // NOLINT(whitespace/braces)
-#endif
-
-extern "C" __aicore__ __attribute__((always_inline)) void kernel_entry(__gm__ int64_t *args) {
-    __gm__ Tensor *src_tensor = reinterpret_cast<__gm__ Tensor *>(args[0]);
-    __gm__ Tensor *out_tensor = reinterpret_cast<__gm__ Tensor *>(args[1]);
-    __gm__ float *src = reinterpret_cast<__gm__ float *>(src_tensor->buffer.addr) + src_tensor->start_offset;
-    __gm__ float *out = reinterpret_cast<__gm__ float *>(out_tensor->buffer.addr) + out_tensor->start_offset;
-    float scalar = from_u64<float>(static_cast<uint64_t>(args[2]));
-
-    constexpr int kRows = 128;
-    constexpr int kCols = 128;
-    using DynShapeDim5 = Shape<1, 1, 1, kRows, kCols>;
-    using DynStrideDim5 = Stride<1, 1, 1, kCols, 1>;
-    using GlobalData = GlobalTensor<float, DynShapeDim5, DynStrideDim5>;
-    using TileData = Tile<TileType::Vec, float, kRows, kCols, BLayout::RowMajor, -1, -1>;
-
-    TileData src_tile(kRows, kCols);
-    TileData dst_tile(kRows, kCols);
-    TASSIGN(src_tile, 0x0);
-    TASSIGN(dst_tile, 0x10000);
-
-    GlobalData src_global(src);
-    GlobalData dst_global(out);
-
-    TLOAD(src_tile, src_global);
-    set_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-    wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-    TADDS(dst_tile, src_tile, scalar);
-    set_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-    wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-    TSTORE(dst_global, dst_tile);
-
-    pipe_sync();
-}
